pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["amrylnikov <amrylnikov@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django-bootstrap5 = "^23.3"
django-filter = "^23.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
